fix circular dependencies
